@@ -3,13 +3,19 @@
   "version": "1.0.0",
   "description": "A simple virtual pet app using Electron",
   "main": "main.js",
+  "type": "module",
   "scripts": {
     "start": "electron .",
-    "dev": "electron . --dev"
+    "dev": "electron . --dev",
+    "lint": "eslint .",
+    "lint:fix": "eslint . --fix"
   },
   "author": "",
   "license": "MIT",
   "devDependencies": {
-    "electron": "^28.0.0"
+    "@eslint/js": "^9.23.0",
+    "electron": "^29.1.0",
+    "eslint": "^9.23.0",
+    "globals": "^16.0.0"
   }
-} 
+}
